backend/requirements.txt
# Web framework and API basics
Flask
Flask-CORS
requests

# AI/ML core libraries
langchain
pydantic
numpy

# Database and persistence
weaviate-client==3.25.3
psycopg2-binary
SQLAlchemy

# Optional dependencies
# torch
# sentence-transformers
